Replace wasmtime-runtime's signal handling
Wasmtime made the internals that we relied on private somewhat recently.
Instead, vendor the parts that we're using and adapt for our needs.

This allows us to - eventually - catch up to upstream cranelift by
reducing the delta. Notably though, in addition to the JIT allocator
patch we need a new cranelift change since upstream recently made trap
codes more compact. We're using many traps per function so we're
essentially reverting this change for now.